rename function to be camelCase
